requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
langchain>=0.1.16
openai>=1.3.5
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
unstructured>=0.11.0
pymupdf>=1.22.0
python-docx>=1.0.0
pdfplumber>=0.10.2
python-dotenv>=1.0.0
python-multipart>=0.0.6
numpy>=1.24.0
pinecone-client>=3.0.0
